[Datalist] Pressing enter without a selected option shouldn't change the input https://bugs.webkit.org/show_bug.cgi?id=189010
Reviewed by Tim Horton.

Source/WebKit:

Currently, the value of an input field gets cleared if there is no selected
datalist suggestion when the enter key is pressed. The correct behavior is to
leave the value of the input as-is.

The incorrect behavior is a consequence of the fact that an empty string is
returned by [WKDataListSuggestionsView currentSelectedString] if there is no
selection. To fix the behavior, the method now returns an std::optional instead
of an empty string. If std::nullopt is returned, we do not make any modification
to the value of the input. This ensures that we can still change the value of
an input field to an empty string in the case that an empty string is part of
the suggestions.

Augmented test: fast/forms/datalist/datalist-textinput-keydown.html

* UIProcess/mac/WebDataListSuggestionsDropdownMac.mm:
(WebKit::WebDataListSuggestionsDropdownMac::selectOption):
(-[WKDataListSuggestionCell drawRect:]): Quick fix. The mouseover color was incorrect.
(-[WKDataListSuggestionsView currentSelectedString]):

LayoutTests:

Augmented test to verify that pressing enter when there is no selected datalist
suggestion does not change the value of the input field.

* fast/forms/datalist/datalist-textinput-keydown-expected.txt:
* fast/forms/datalist/datalist-textinput-keydown.html:


Identifier: 204081@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@235427 268f45cc-cd09-0410-ab3c-d52691b4dbfc