Unreviewed, fix incorrect WPE build fix made in r236009
This isn't the right place to add the typedef. I failed to notice that the WebKitWebView
type is not actually used anywhere in this header. Actually, the problematic function is no
longer declared or used anywhere and was just missed when WebKitScriptDialog was refactored,
so remove it.

* UIProcess/API/wpe/WebKitScriptDialog.h:
* UIProcess/API/wpe/WebKitScriptDialogWPE.cpp:
(webkitScriptDialogRun): Deleted.

Identifier: 204618@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@236051 268f45cc-cd09-0410-ab3c-d52691b4dbfc